[SDK-109] JMS communication has been outsourced to its own module in the project
