+++ njams-sdk-communication-jms/.gitignore
@@ -1,0 +1,4 @@
+/target/
+/.classpath
+/.project
+/njams4-sdk-communication-jms.iml
