+++ njams-sdk-communication-jms/src/main/java/com/im/njams/sdk/communication/jms/connector/JmsConnector.java
@@ -1,37 +1,16 @@
 /*
- * Copyright (c) 2019 Faiz & Siegeln Software GmbH
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"),
- * to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense,
- * and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
- *
- * The Software shall be used for Good, not Evil.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
- * IN THE SOFTWARE.
  */
 
-package com.im.njams.sdk.communication.jms;
+package com.im.njams.sdk.communication.jms.connector;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.im.njams.sdk.common.JsonSerializerFactory;
 import com.im.njams.sdk.common.NjamsSdkRuntimeException;
-import com.im.njams.sdk.communication.connection.AbstractConnector;
+import com.im.njams.sdk.communication.connector.AbstractConnector;
+import com.im.njams.sdk.communication.jms.JmsConstants;
 import com.im.njams.sdk.settings.PropertyUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.jms.Connection;
-import javax.jms.ExceptionListener;
-import javax.jms.Session;
-import javax.jms.JMSException;
-import javax.jms.ConnectionFactory;
-import javax.jms.JMSContext;
-
+import javax.jms.*;
 import javax.naming.InitialContext;
 import javax.naming.NamingException;
 import java.util.*;
@@ -40,7 +19,7 @@ import java.util.*;
  * JMS implementation for Receiver and Sender
  *
  * @author krautenberg@integrationmatters.ocm
- * @version 4.0.6
+ * @version 4.1.0
  */
 public abstract class JmsConnector extends AbstractConnector implements ExceptionListener {
