+++ njams-sdk-communication-jms/src/main/resources/services/com.im.njams.sdk.communication.connectable.Receiver
@@ -1,0 +1,1 @@
+com.im.njams.sdk.communication.jms.connectable.JmsReceiver
