+++ njams-sdk/pom.xml
@@ -139,7 +139,7 @@
         <dependency>
             <groupId>com.faizsiegeln</groupId>
             <artifactId>njams-messageformat</artifactId>
-            <version>4.0.1-SNAPSHOT</version>
+            <version>4.0.1</version>
         </dependency>
 
         <dependency>
@@ -174,13 +174,6 @@
             <scope>compile</scope>
         </dependency>
 
-        <!-- provided - not shaded -->
-        <dependency>
-            <groupId>javax</groupId>
-            <artifactId>jms</artifactId>
-            <version>2.0</version>
-            <scope>provided</scope>
-        </dependency>
         <dependency>
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-api</artifactId>
